Added multithreading
To run this on you local instance, first, use a mac. Second, if you run
in eclipse you should get the console to prompt you for input. Select
'q' then enter 'ERROR' as your query. I only have this running on 1
client thread for now. Just change the for loop to N number of servers.