Preliminary UDP group membership (NOT TESTED)
Started basic goup management specifically around a node joining the
group. The idea is to have a synchronized array list in the main app
and then update this from the one local thread (group server) that is
listening/managing join/drop requests. LINUX_7 is the contact node
(head of the array list) so all nodes will have this hard coded as the
node to join through.