Goup Multicasting
- Multiple nodes can now join the group
- When this happens, they are added to the groupmembershiplist on
Linux5 (the introducer) and then linux5 sends out a multicast to all
nodes listening on group 228.5.6.7
- Added more conversational logging (Feel free to delete, i was bored)
- Currently the update message only contains "HIIII". This will need to
be changed to something like: "R:<TIMESTAMP>:<IP>:<PORT>" for removing
a node and the "A:<TIMES…." for adding a node (we can't ship the whole
list around because we're restricted to byte[256] array in size so it
wouldn't scale. We'll just have to send an Add/remove message.
- Next steps would be to get a protocol class in place to handle this
parsing.